add restcontroller and getmapping

+++ src/main/java/com/example/quotes/repositories/QuoteRepository.java
@@ -3,5 +3,5 @@ package com.example.quotes.repositories;
 import com.example.quotes.models.Quote;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuoteRepository extends JpaRepository<Quote, int> {
+public interface QuoteRepository extends JpaRepository<Quote, Long> {
 }
